backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy>=2.0.36
psycopg[binary]
alembic==1.12.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator==2.1.0
dnspython>=2.0.0
aiofiles==23.2.1
Pillow>=10.1.0
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
opencv-python>=4.8.0
numpy>=1.26.0
openai>=1.12.0
python-docx==1.1.0
reportlab==4.0.7
pdfkit==1.0.0
xhtml2pdf==0.2.15
playwright>=1.40.0
qrcode==7.4.2
jinja2==3.1.2
aiosmtplib==3.0.1
cryptography==41.0.7
ultralytics>=8.2.0

